ERA-8825 User etag calculation does not include linked subject (#2714)
* ERA-8825 User etag calculation does not include linked subject

fix: added the user linked subject id and user linked subject updated_at the hash for a user. This tracks adding/removing user linked subject and if the subject itself changed. This of course is done for any user's child profiles.

* fix: address PR comments, restructure extracting data out of the user for calculating the etag
fix: update test_additional assert when comparing to see if a string is empty. Want to use "==" over "is".
fix: import Subject over using the django content type system.

---------

Co-authored-by: “Chris <“chrisj@earthranger.com”>