ERA-8891 Update rt_api client class to support tenants (#2747)
feat: add tenant id and domain to ClientData, refactor ClientData to use a frozen dataclass
feat: add the TenantContextManager to the authentication stage in the rt_api.
fix: update some unit tests, waiting for the subject tenant_id+id index to be fixed before this test will pass
fix: remove the need to start the rt_api wsgi server with the default tenant
feat: for me, added a DjangoSettingsClient() so that I can load Django Settings as TenantSettings, just for testing.

Co-authored-by: “Chris <“chrisj@earthranger.com”>