ERA-8901 500 from api/v1.0/status when getting status for a tenant (#2763)
feat: UnsetDASTenantContextManager to disable the current_tenant set on the local thread, so that we can query for all records in the DASTenant model. Without being filtered down to the single current tenant.
fix: patch loaddata_with_tenant until we get ERA-8886 merged. It simply sets the tenant to FQDN
fix: two signals in activity.signals that assumed no tenant set meant to use the FQDN. We should throw an exception and fix the code rather than do something to a hardcoded tenant.

Co-authored-by: “Chris <“chrisj@earthranger.com”>