ERA-10316 Cleanup Errors that are not Errors in production (#3219)
* ERA-10316 Cleanup Errors that are not Errors in production
fix: use graceful celery queue for the analyze_subject celery task
fix: handle a list being returned in data when we are logging detailed information on a  400 or above request status result. This is to help us better troubleshoot clients.
fix: create a log filter and apply so that we can add the tenant domain name in all of our log output. Needed to cleanup the utils.tenant.__init__.py imports to prevent loading of some of the tenant code before django had initialized.

* fix: the OverAllTenantTask does inherit from QueueOnce, so any task using this, should set graceful true, otherwise we see spurious errors in the logs. Especially the broadcast service status task.

---------

Co-authored-by: chrisj-er <“chrisj@earthranger.com”>