ERA-8935 Server 500 from sensors API duplicate keys in latestobservationSource table (#3349)
* ERA-8935 Server 500 from sensors API duplicate keys in latestobservationSource table
fix: race condition in the insert logic when data is being pushed to the sensors api on mulitple threads. Added On Conflict to the insert, and will update the existing record in the On Action.
fix: catch IntegrityErrors, returning those as a 409 conflict with any other errors in the sensors handler

* feat: for single tenant sites, increase the API workers to 3, add graceful timeout of 40 seconds and a keepalive of 10 seconds. Default keepalive is 2.

---------

Co-authored-by: “chrisj-er” <“chrisj@earthranger.com”>